Move irc-related classes to sub-package
